@@ -1,32 +1,34 @@
 {
-  "name": "dash",
-  "version": "0.1.0",
+  "name": "test-dashboard",
+  "version": "0.0.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite --host",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
   },
   "dependencies": {
-    "apexcharts": "^3.36.0",
-    "axios": "^1.1.2",
-    "chart.js": "^2.9.4",
-    "core-js": "^3.8.3",
-    "jsonwebtoken": "^8.5.1",
-    "vue": "^2.6.14",
+    "@tailwindcss/forms": "^0.5.3",
+    "axios": "^1.2.2",
+    "material-icons": "^1.13.1",
+    "pinia": "^2.0.28",
+    "prettier-plugin-tailwindcss": "^0.2.1",
+    "sass": "^1.57.1",
+    "vue": "^3.2.45",
     "vue-apexcharts": "^1.6.2",
-    "vue-router": "^3.5.1",
-    "vue-toasted": "^1.1.28",
-    "vuex": "^3.6.2",
-    "vuex-persistedstate": "^4.1.0"
+    "vue-m-message": "^4.0.1",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-plugin-vuex": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "node-sass": "^8.0.0",
-    "sass-loader": "^13.2.0",
-    "vue-template-compiler": "^2.6.14"
+    "@vitejs/plugin-vue": "^4.0.0",
+    "autoprefixer": "^10.4.13",
+    "eslint": "^8.22.0",
+    "eslint-plugin-vue": "^9.3.0",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.2",
+    "tailwindcss": "^3.2.4",
+    "vite": "^4.0.0"
   },
   "browserslist": [
     "> 1%",
